LibArgParse: Added argparser library
This argparser is heavily inspired by the one used on the radare2 tools.

It is configured the same way, with a string you pass to the
initializer. In this string you specify the options that require an
argument.

One feature I could improve on the future is that you could pass a list
of arguments to a given option.